+++ py2app_tests/pkg_script_app/quot.py
@@ -12,9 +12,7 @@ def import_module(name):
         print ("* import failed")
 
     else:
-        for k in name.split('.')[1:]:
-            m = getattr(m, k)
-        print (m.__name__)
+        print m.__name__
 
 def print_path():
     print(sys.path)